.requirements.txt
numpy==1.26.4
opencv-python==4.7.0.72
scipy==1.10.1
scikit-image==0.21.0
pillow==12.0.0
matplotlib